Rename and move components to appropriate packages
