+++ client/src/app.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import NavigationBar from './components/navigation-bar';
+import NavigationBar from './components/shared/navigation/navigation-bar';
 import Home from './pages/home';
 import NoMatch from './pages/no-match';
-import routeData from './components/data/navigation-route';
+import routeData from './components/data/route-data';
 import RouteComponent from './components/shared/navigation/navigation-route';
 
 class App extends Component {
