+++ client/src/components/shared/button/reset-button.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import SearchButton from './search/search-button';
+import SearchButton from './search-button';
 
 class ResetButton extends Component {
   render() {
