+++ client/src/components/shared/event/event-search-form.js
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import SearchForm from '../search/search-form';
+
+class EventSearchForm extends React.Component {
+  render() {
+    return (
+      <SearchForm placeholder="Search event" />
+    );
+  }
+}
+
+export default EventSearchForm;
