--- client/src/components/shared/search/local-search.js
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import SearchForm from '../search-form';
-
-class LocalSearch extends React.Component {
-  render() {
-    return (
-      <SearchForm placeholder="Search event" />
-    );
-  }
-}
-
-export default LocalSearch;
