+++ client/src/components/shared/search/search-filter-toolbar.js
@@ -1,16 +1,16 @@
 import React, { Component } from 'react';
 import { ButtonToolbar } from 'react-bootstrap';
 import PropTypes from 'prop-types';
-import DropdownWithDate from '../event-search/custom-dropdown';
-import ResetButton from './reset-button';
-import DropdownMaker from './dropdown/dropdown-maker';
+import DropdownWithDate from '../../event-search/custom-dropdown';
+import ResetButton from '../button/reset-button';
+import DropdownFactory from '../dropdown/dropdown-factory';
 
 class SearchFilterToolbar extends Component {
   render() {
     const { data } = this.props;
     return (
       <ButtonToolbar className="justify-content-center">
-        <DropdownMaker data={data} />
+        <DropdownFactory data={data} />
         <DropdownWithDate />
         <ResetButton />
       </ButtonToolbar>
